+++ src/darker/tests/test_main.py
@@ -1,3 +1,8 @@
+"""Unit tests for :mod:`darker.__main__`"""
+
+import logging
+import os
+import re
 from pathlib import Path
 from subprocess import check_call
 from types import SimpleNamespace
@@ -10,6 +15,7 @@ import darker.__main__
 import darker.import_sorting
 from darker.git import RevisionRange
 from darker.utils import TextDocument
+from darker.verification import NotEquivalentError
 
 
 def test_isort_option_without_isort(tmpdir, without_isort, caplog):
@@ -156,6 +162,43 @@ def test_format_edited_parts_all_unchanged(git_repo, monkeypatch):
     assert result == []
 
 
+def test_format_edited_parts_ast_changed(git_repo, caplog):
+    """``darker.__main__.format_edited_parts()`` when reformatting changes the AST"""
+    caplog.set_level(logging.DEBUG, logger="darker.__main__")
+    paths = git_repo.add({"a.py": "1\n2\n3\n4\n5\n6\n7\n8\n"}, commit="Initial commit")
+    paths["a.py"].write_bytes(b"8\n7\n6\n5\n4\n3\n2\n1\n")
+    with patch.object(
+        darker.__main__, "verify_ast_unchanged"
+    ) as verify_ast_unchanged, pytest.raises(NotEquivalentError):
+        verify_ast_unchanged.side_effect = NotEquivalentError
+
+        _ = list(
+            darker.__main__.format_edited_parts(
+                git_repo.root,
+                [Path("a.py")],
+                RevisionRange("HEAD"),
+                enable_isort=False,
+                black_args={},
+            )
+        )
+    a_py = str(paths["a.py"])
+    main = "darker.__main__:__main__.py"
+    log = [
+        line
+        for line in re.sub(r":\d+", "", caplog.text).splitlines()
+        if " lines of context " in line
+    ]
+    assert log == [
+        f"DEBUG    {main} AST verification of {a_py} with 0 lines of context failed",
+        f"DEBUG    {main} Trying with 5 lines of context for `git diff -U {a_py}`",
+        f"DEBUG    {main} AST verification of {a_py} with 5 lines of context failed",
+        f"DEBUG    {main} Trying with 7 lines of context for `git diff -U {a_py}`",
+        f"DEBUG    {main} AST verification of {a_py} with 7 lines of context failed",
+        f"DEBUG    {main} Trying with 8 lines of context for `git diff -U {a_py}`",
+        f"DEBUG    {main} AST verification of {a_py} with 8 lines of context failed",
+    ]
+
+
 @pytest.mark.kwparametrize(
     dict(
         arguments=["--diff"],
@@ -224,7 +267,7 @@ def test_main(
     retval = darker.__main__.main(arguments + ['a.py'])
 
     stdout = capsys.readouterr().out.replace(str(git_repo.root), '')
-    assert stdout.split("\n") == expect_stdout
+    assert stdout.replace(os.sep, "/").split("\n") == expect_stdout
     assert paths["a.py"].read_bytes().decode("ascii") == newline.join(expect_a_py)
     assert paths["b.py"].read_bytes().decode("ascii") == "print(42 ){newline}"
     assert retval == expect_retval